refactor: Set default values for form elements in abrsd_user_registration.module
The `abrsd_user_registration.module` file has been modified to set default values for form elements. This change introduces a function `setDefaultValue` that iterates through a list of form element keys and sets the default value based on submission data or user field value. This improves the code by centralizing the logic for setting default values and makes it easier to maintain and update the form elements in the future.