refactor: Hide blog tags description content in taxonomy term view
The `_views.scss` file in the `abrsd` theme has been updated to hide the blog tags in the taxonomy term view. This change adds a new CSS rule to the `.view-taxonomy-term .vocabulary-blog-tags` selector, setting the `display` property to `none`. This improves the visual presentation of the taxonomy term view by removing the display of blog tags description content.