refactor: Update UserRegistrationHelper.php to handle new password in updateUserAccount()
The UserRegistrationHelper.php file has been modified to handle the case where a new password is provided in the updateUserAccount() method. This change adds a conditional check to see if the 'new_password' value is not empty, and if so, it sets the new password for the user using the setPassword() method. This improvement ensures that the user's password is updated when a new password is provided during the account update process.