refactor: Add link to password reset form in user login form
The abrsd_user_registration.module file has been modified to add a link to the password reset form in the user login form. This change adds a new markup element to the form, which generates a link using the Link::fromTextAndUrl() method. The link is positioned correctly using the '#weight' property. This improvement enhances the user experience by providing an easy way for users to reset their passwords when logging in.