refactor: Add container template for user password reset form
The code changes in this commit add a new container template, `container--user-pass-reset.html.twig`, for the user password reset form. This template is used to wrap the child elements of the form and applies specific HTML attributes and CSS classes. This improvement enhances the layout and styling of the user password reset form.