refactor: Update page--node--blog.html.twig to display the post label, comment count, and date posted in the floating title banner.
This commit updates the page--node--blog.html.twig template file to include the post label, comment count, and date posted. The post label is set to 'Page Title' if it is empty. The comment count is retrieved from the node's field_blog_comment field. The date posted is formatted using the node's created time. This change improves the display of the blog page by providing more information about each blog post.