refactor: Rearrange the order of elements in the blog page title banner
This commit rearranges the order of elements in the page--node--blog.html.twig template file to display the post label, comment count, and date posted in the floating title banner. The post label is set to 'Page Title' if it is empty. The comment count is retrieved from the node's field_blog_comment field. The date posted is formatted using the node's created time. This change improves the visual hierarchy and readability of the blog page.