refactor: Include webform template for user registration form
This commit adds a new webform template, webform--user-registration.html.twig, to the abrsd theme. The template is a copy of the webform.html.twig theme_wrapper and includes the necessary variables for contextual links to appear. This change improves the theming of the user registration form by providing a customized template specifically for this form.