refactor: Update form submission loader to include updateTime function
This commit updates the form submission loader in the abrsd theme to include a new function called updateTime. This function is called when the form submission button is clicked and retrieves the serverTime value from the drupalSettings object. It then updates the form message element with the registration form creation date and time. This change improves the functionality of the form submission loader by displaying the accurate creation date and time for the Registration form.