refactor: Update 'new_password' element data to use translation function
This commit updates the 'new_password' element data in the UserRegistration.php file to use the translation function. Instead of hardcoding the string 'Password changed by user', it now uses the translation function $this->t('Password changed by user'). This change improves the internationalization and localization capabilities of the code.