refactor: Update gulpfile.js and package.json for vanilla-cookieconsent integration
This commit updates the gulpfile.js and package.json files in the abrsd theme to integrate the vanilla-cookieconsent library. The gulpfile.js is modified to include the vanilla-cookieconsent JavaScript file in the list of files to be moved to the js folder. The package.json is updated to add the vanilla-cookieconsent dependency. This change enhances the functionality of the theme by adding support for cookie consent management.