refactor: Update link text handling in banner and about us templates (a11y).
This commit removes the uppercase transformation of the link text in the banner and about us templates. Previously, if no link text was provided, the default value 'Read More' would be transformed to uppercase. This commit removes that transformation, allowing the link text to be displayed as-is. This improves consistency and flexibility in the templates.