refactor: Update _hero-cta-floater.html.twig to remove uppercase transformation of link text (a11y).
This commit removes the uppercase transformation of the link text in the _hero-cta-floater.html.twig template. Previously, if no link text was provided, the default value 'Read More' would be transformed to uppercase. This commit removes that transformation, allowing the link text to be displayed as-is. This improves consistency and flexibility in the template.