refactor: Remove uppercase transformation of link text in resource cards grid template (a11y)
This commit removes the uppercase transformation of the link text in the resource cards grid template. Previously, if no link text was provided, the default value 'Read More' would be transformed to uppercase. This commit removes that transformation, allowing the link text to be displayed as-is. This improves consistency and flexibility in the template.