feat: Update user docs modal with larger size
The code changes in this commit modify the user docs modal in the `page--node--user--create-member-account.html.twig` template. The modal dialog has been updated to use a larger size by adding the `modal-lg` class to the modal dialog container. This change improves the user experience by providing more space for displaying the modal content.