feat: Add data attribute with node URL to user docs modal
The code changes in this commit add a data attribute to the user docs modal in the `abrsd_user_docs.js` file. This data attribute stores the path of the node associated with the modal content. Additionally, an event listener is added to the modal to execute code after the modal is closed. This enhancement improves the functionality and usability of the user docs modal.