refactor: Prefill user docs modal content and title
The code changes in this commit update the content and title of the user docs modal in the `page--node--user--create-member-account.html.twig` template. The modal now displays the ABR & SD Privacy Policy, Terms of Use, and other legal documents. The modal title has been changed to "ABR & SD Document Title" for clarity. This enhancement improves the user experience by providing relevant and informative content in the modal.