refactor: Disable 'Accept' button in user docs modal
The code changes in this commit disable the 'Accept' button in the user docs modal in the `page--node--user--create-member-account.html.twig` template. The button is now disabled by adding the `disabled` class to the button element. This refactor aligns the button state with the intended functionality of the modal, preventing users from accepting the terms without reviewing the content.