refactor: Set terms_of_use and code_of_conduct fields to 1 in preSave()
The code changes in this commit set the terms_of_use and code_of_conduct fields to 1 in the preSave() method of the UserRegistration.php file. This ensures that the corresponding form fields are checked by the user. This refactor improves the accuracy of user data and aligns with the intended functionality of the form.