refactor: Update code to add scrollThreshold property to abrsdUserDocs behavior
The code changes in this commit add a scrollThreshold property to the abrsdUserDocs behavior in the abrsd_user_docs.js file. This property is used to determine the scroll threshold for enabling the 'Accept' and 'Reject' buttons in the user docs modal. By making this change, the scroll threshold can be easily adjusted in the future if needed, improving the flexibility and maintainability of the code.