refactor: Improve modal functionality and code organization. Expand the click event surface to "div.modal-conent" so that the 'close' button click can be tracked.
This commit refactors the abrsd_user_docs.js file to improve the functionality and organization of the modal feature. It updates the code to use the behaviors.abrsdUserDocs.docsModal property instead of querying the modal element multiple times. Additionally, it improves the loading message display and disables modal buttons when necessary. This change enhances the overall user experience and maintainability of the code.