refactor: Add blog list sidebar top card template and rename views row style template.
This commit adds a new template file, _blog_list_sidebar_top_card.html.twig, to the abrsd theme's templates/partials directory. The template is used to display a card in the blog list in the sidebar. It includes a title, a description, and a "See All Blogs" button. This addition enhances the visual presentation and user experience of the blog list in the sidebar.